requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
python-dotenv
sendgrid
twilio